попытка доделать
