+++ index.html
@@ -117,6 +117,7 @@
                 <fieldset class="img-upload__text text">
                   <div class="img-upload__field-wrapper">
                     <input class="text__hashtags" name="hashtags" type="text" placeholder="#ХэшТег">
+                    <span>Ошибка</span>
                   </div>
                   <div class="img-upload__field-wrapper">
                     <textarea class="text__description" name="description" placeholder="Ваш комментарий..." maxlength="140"></textarea>
