+++ js/main.js
@@ -2,6 +2,7 @@ import * as data from './data.js';
 import {createThumbnails} from './render-thumbnails.js';
 import {showPopupData, closePopupData} from './popup-thumbnail.js';
 import {isEscapeKey} from './util.js';
+import {validateHashTeg} from './validate.js';
 
 const COUNT_OBJECT = 25;
 
@@ -19,22 +20,22 @@ const buttonUpload = document.querySelector('.img-upload__input');
 const imgUpload = document.querySelector('.img-upload__overlay');
 const imgUploadCancel = document.querySelector('.img-upload__cancel');
 
-const onPopupEscKeydown = (evt) => {
+function onPopupEscKeydown(evt){
   if (isEscapeKey(evt)) {
     evt.preventDefault();
     body.classList.remove('modal-open');
     closePopupData();
   }
-};
+}
 
-const onCloseEscKeydown = (evt) => {
+function onCloseEscKeydown (evt){
   if (isEscapeKey(evt)) {
     evt.preventDefault();
     body.classList.remove('modal-open');
     imgUpload.classList.add('hidden');
     buttonUpload.value = '';
   }
-};
+}
 
 const textHashtags = document.querySelector('.text__hashtags');
 
@@ -71,36 +72,9 @@ imgForm.addEventListener('submit', (evt) => {
   evt.preventDefault();
   const isValid = pristine.validate();
   validateHashTeg(textHashtags.value);
-  //валидация хеш тега
   if (isValid) {
     console.log('все норм');
   } else {
     console.log('не норм');
   }
 });
-
-function validateHashTeg(hashteg){
-  hashteg = hashteg.split(' ');
-  if(hashteg.length>5){
-    return 'нельзя указывать больше 5 хештегов';
-  }
-  const allowCharacters = /(([a-zA-Z].*\d)|(\d.*[a-zA-Z]))/;
-  // строка после решётки должна состоять из букв и чисел и не может содержать пробелы, спецсимволы (#, @, $ и т. п.), символы пунктуации (тире, дефис, запятая и т. п.), эмодзи и т. д.;
-  for (let i = 0; i<hashteg.length;i++){
-    if(hashteg[i][0] !== '#'){
-      return 'хэш-тег начинается с символа #';
-    }
-    if (allowCharacters.test(hashteg[i]) === false){
-      return 'строка после решётки должна состоять из букв';
-    }
-    if(hashteg[i].length>20){
-      return 'максимальная длина одного хэш-тега 20 символов';
-    }
-    if(hashteg[i]==='#'){
-      return 'хеш-тег не может состоять только из одной решётки';
-    }
-    if(hashteg.indexOf(hashteg[i]) !== -1){
-      return 'один и тот же хэш-тег не может быть использован дважды';
-    }
-  }
-}
